Allowing faster graceful shutdown. Allowing faster services startup.
