+++ gradle.properties
@@ -1,2 +1,2 @@
-version=1.38.0
+version=1.39.0
 org.gradle.internal.http.socketTimeout=120000
