+++ integration-tests/src/test/java/com/transferwise/tasks/testapp/dao/JdbcTaskDaoIntTest.java
@@ -1,0 +1,5 @@
+package com.transferwise.tasks.testapp.dao;
+
+class JdbcTaskDaoIntTest extends TaskDaoIntTest {
+
+}
