--- integration-tests/src/test/java/com/transferwise/tasks/testapp/dao/MySqlTaskDaoIntTest.java
@@ -1,5 +1,0 @@
-package com.transferwise.tasks.testapp.dao;
-
-class MySqlTaskDaoIntTest extends TaskDaoIntTest {
-
-}
