+++ tw-tasks-core/src/main/java/com/transferwise/tasks/EnvironmentValidator.java
@@ -1,6 +1,8 @@
 package com.transferwise.tasks;
 
 import com.vdurmont.semver4j.Semver;
+import javax.validation.ConstraintViolationException;
+import javax.validation.Validator;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -9,9 +11,17 @@ public class EnvironmentValidator implements IEnvironmentValidator {
   @Autowired
   private TasksProperties tasksProperties;
 
+  @Autowired
+  private Validator validator;
+
   @Override
   public void validate() {
-    String previousVersion = tasksProperties.getEnvironment().getPreviousVersion();
+    var violations = validator.validate(tasksProperties);
+    if (!violations.isEmpty()) {
+      throw new ConstraintViolationException(violations);
+    }
+
+    var previousVersion = tasksProperties.getEnvironment().getPreviousVersion();
     if (StringUtils.trimToNull(previousVersion) == null) {
       throw new IllegalStateException("Previous version has not been specified.");
     }
