+++ tw-tasks-core/src/main/java/com/transferwise/tasks/config/ResolvedValue.java
@@ -6,6 +6,8 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+import java.util.Collection;
+import java.util.Map;
 import javax.validation.Constraint;
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
@@ -23,15 +25,42 @@ public @interface ResolvedValue {
 
   Class<? extends Payload>[] payload() default {};
 
-  class ResolvedValueValidator implements ConstraintValidator<ResolvedValue, String> {
+  class ResolvedValueValidator implements ConstraintValidator<ResolvedValue, Object> {
 
     @Override
     public void initialize(ResolvedValue contactNumber) {
+      // Nothing to initialize
     }
 
     @Override
-    public boolean isValid(String field, ConstraintValidatorContext ctx) {
-      return field == null || !field.contains("${");
+    @SuppressWarnings("unchecked")
+    public boolean isValid(Object field, ConstraintValidatorContext ctx) {
+      if (field instanceof String) {
+        return isValidString((String) field);
+      } else if (field instanceof Map) {
+        var map = (Map<?, ?>) field;
+        for (var entry : map.entrySet()) {
+          var key = entry.getKey();
+          if (key instanceof String) {
+            return isValidString((String) key);
+          }
+          var value = entry.getValue();
+          if (value instanceof String) {
+            return isValidString((String) value);
+          }
+        }
+      } else if (field instanceof Collection) {
+        for (var val : (Collection<?>) field) {
+          if (val instanceof String) {
+            return isValidString((String) val);
+          }
+        }
+      }
+      return true;
+    }
+
+    protected boolean isValidString(String val) {
+      return val == null || !val.contains("${");
     }
   }
 }
