+++ tw-tasks-core/src/main/java/com/transferwise/tasks/processing/TasksProcessingService.java
@@ -699,6 +699,11 @@ public class TasksProcessingService implements GracefulShutdownStrategy, ITasksP
     shutdownStartTime = Instant.now(TwContextClockHolder.getClock());
     shuttingDown = true;
     tasksProcessingExecutor.shutdown();
+    for (var bucketId : bucketsManager.getBucketIds()) {
+      var bucket = globalProcessingState.getBuckets().get(bucketId);
+      // Wakes up waiting threads.
+      bucket.increaseVersion();
+    }
   }
 
   @Override
